src/core/requirements.txt
amqp==5.0.9
anyio==3.5.0
appdirs==1.4.3
asgiref==3.5.0
attrs==21.4.0
Authlib==1.0.0
bcrypt==3.2.0
bidict==0.21.2
billiard==3.6.4.0
borgbackup==1.1.17
CacheControl==0.12.6
celery==5.2.1
celery-once==3.0.1
celery-redbeat==2.0.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
contextlib2==0.6.0
coverage==6.3.1
croniter==1.3.4
cryptography==3.2.1
cs==3.0.0
distlib==0.3.0
distro==1.4.0
dnspython==1.16.0
ecdsa==0.17.0
eventlet==0.33.0
fastapi==0.75.0
flower==1.0.0
graypy==2.1.0
greenlet==1.1.2
h11==0.12.0
html5lib==1.0.1
httpcore==0.14.7
httpx==0.22.0
humanize==3.2.0
idna==2.10
importlib-metadata==3.3.0
iniconfig==1.1.1
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.2.2
libvirt-python==6.10.0
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
mysql-connector-python==8.0.22
packaging==20.3
paramiko==2.7.2
pep517==0.8.2
pluggy==1.0.0
progress==1.5
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.14.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.9.0
PyJWT==2.3.0
PyNaCl==1.4.0
pyparsing==2.4.6
pytest==7.0.1
pytest-cov==3.0.0
python-dateutil==2.8.1
python-engineio==4.0.0
python-jose==3.3.0
pytoml==0.1.21
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.25.1
retrying==1.3.3
rfc3986==1.5.0
rsa==4.8
six==1.15.0
slack-sdk==3.4.2
sniffio==1.2.0
SQLAlchemy==1.4.32
sqlalchemy2-stubs==0.0.2a20
sqlmodel==0.0.6
starlette==0.17.1
tenacity==7.0.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.17.6
vine==5.0.0
waitress==1.4.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
yml==0.0.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
